hélt aðeins að vinna í þessu.

+++ mqtt_client_CC3220SF_LAUNCHXL_tirtos_gcc/client_cbs.h
@@ -82,7 +82,7 @@ struct client_info
 
 struct msgQueue
 {
-    float event;
+    int event;
     void        *msgPtr;
     int32_t topLen;
 };
